F6 · Saneo del nucleo y de los scripts de auditoria
Refactor puro: ningun numero se mueve. La linea base lo vigila, y ademas se
verifico corriendo `auditoria_transporte.py` antes y despues — misma fila
(auto 16,95 · metro 32,79 · bici 22,84 · caminata 7,98 · v/c 1,38 · f_op 5,9).

FORWARDING DE OFERTA, unificado. Llamar a las tres funciones de oferta exige
desarmar la configuracion campo por campo: ~43 lineas de reenvio que estaban
escritas TRES veces (el loop del MSA, la red vacia de coupled_metrics, y el
script que audita el todo-o-nada). No era inocente: al retirar `tasa_carga` en
F3 hubo que tocar los tres sitios, y el que vive en un script no lo cubre
ningun test — si se hubiera olvidado, habria fallado recien al correrlo a mano.
Ahora es `supply/oferta.py::resolver_oferta`, mas `resolver_red_vacia` para el
contrafactual del modulo acoplado.

SCRIPTS. Los once compartian el arranque —armar la config de la app, drenar el
MSA, resumir el reparto— reescrito ocho veces, con la expresion del v/c copiada
textual en cinco. Y de los cinco dialectos de "mover un parametro", UNO solo
llevaba el guard que atrapa el fallo silencioso de `model_copy(update=...)`.
Ahora `scripts/_comun.py` lo declara una vez y TODOS los `con_*` validan.

De paso desaparecen los ocho `sys.path.insert`: eran innecesarios, Python ya
pone el directorio del script en el path.

EL COMENTARIO FALSO. `sensibilidad.py` afirmaba que la paridad de los betas con
el espejo TS estaba "verificada por el test de contrato". Era falso —el golden
cubria city, supply, globales, sim y land_use, y los betas quedaban afuera, que
es justo el bloque que habia driftado—. Reescrito con lo que pasa hoy: el lado
TS se genera desde el mismo dict, asi que no hay paridad que verificar.

RUFF, que ahora significa algo. El CI corre `ruff check` desde F4, pero el
paquete emitia ~100 avisos: casi todos por notacion matematica (α, β, ρ, − en
docstrings; variables S, L, N, Q, H; funciones como `_aggregate_T_expected`).
Un lint siempre rojo no lo lee nadie. Se declararon esas familias como lo que
son —notacion del modelo, que reemplazada por ASCII haria mas dificil cotejar
el codigo con el Overleaf— y se arreglaron los avisos REALES que quedaban
debajo: zip sin strict, dict() innecesarios, lineas largas, casts redundantes.
`ruff check` pasa limpio.

Excepcion deliberada: los `l` de dos tests SI se renombraron. `I` como numero
de celdas es notacion; `l` se confunde con 1 y no defiende nada.

apps/api/tests entra en `testpaths` y gana `__init__.py`, como el core: sin eso
`pytest` desde la raiz del monorepo no los recogia.

Verificado: 88 tests del core, 4 de la api, ruff limpio en src/tests/tools/
scripts, typecheck, 58 e2e, formato, y los scripts de auditoria corriendo de
punta a punta (auditoria_transporte, sensibilidad, paradojas,
diagnostico_calibracion, diagnostico_elasticidades, auditoria_suelo).

+++ packages/titirilquen_core/src/titirilquen_core/population.py
@@ -55,7 +55,7 @@ def generar_poblacion(
         [i for i in range(ciudad.n_celdas) if i != ciudad.cbd_index], dtype=np.int64
     )
     objetivo_por_celda = densidad_hab_km * ciudad.ancho_celda_km
-    total = int(round(objetivo_por_celda * celdas_validas.size))
+    total = round(objetivo_por_celda * celdas_validas.size)
     if total <= 0:
         return []
     conteo_por_celda = _mayor_residuo(np.full(celdas_validas.size, objetivo_por_celda), total)
@@ -150,9 +150,9 @@ def generar_poblacion_desde_land_use_det(
             estrato: StratumId = estratos_valid[h]
             s = demand_config.estratos[estrato]
             prob_tele = min(1.0, s.prob_teletrabajo * teletrabajo_factor)
-            n_tele = int(round(n * prob_tele))
+            n_tele = round(n * prob_tele)
             n_work = n - n_tele
-            n_auto = int(round(n_work * s.prob_auto))
+            n_auto = round(n_work * s.prob_auto)
             for k in range(n):
                 tele = k < n_tele
                 # entre los trabajadores, los primeros `n_auto` tienen auto
